requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
pyyaml>=5.4.0
pytest>=6.0.0

# Development dependencies
black
flake8
mypy
pytest-cov
sphinx
sphinx-rtd-theme 